pyproject.toml
[tool.poetry]
name = "alchemy-practice"
version = "0.1.0"
description = ""
authors = ["EugeneSpiridonov <eugenespint@gmail.com>"]
readme = "README.md"
packages = [{include = "alchemy_practice"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
sqlalchemy = "^2.0.23"
alembic = "^1.13.0"
asyncpg = "^0.29.0"
jinja2 = "^3.1.2"
pydantic-settings = "^2.1.0"
sqladmin = "^0.16.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
